Added SQLConnection class, cleaned up testing pages.
SQLConnection class makes it so you only have to store SQL credentials
in one place. Its primary function is to make it easy to transfer code
from local to remote servers with different SQL server credentials. It
also makes all the other code cleaner and easier to use.

Also of note is the addition (in the last commit) of the view page.
Right now, this shows only author and text of the monologue with no
style, so it should be edited later. Basic functionality is working
though.

Total time: 10 hrs